requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Django==3.2.6
django-cas-ng==4.2.1
django-graphql-jwt==0.3.3
django-extensions==3.1.5
Faker==8.11.0
fastdiff==0.3.0
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
lxml==4.6.3
mixer==7.1.2
msgpack==1.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.3.1
promise==2.3
psycopg2-binary==2.8.6
PyJWT==2.1.0
python-cas==1.5.0
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.3
requests==2.26.0
Rx==1.6.1
singledispatch==3.6.2
six==1.16.0
snapshottest==0.6.0
sqlparse==0.4.1
termcolor==1.1.0
text-unidecode==1.3
tomli==1.2.1
urllib3==1.26.6
wasmer==1.0.0
wasmer-compiler-cranelift==1.0.0
